+++ Modules/cjkcodecs/_codecs_iso2022.c
@@ -2,6 +2,7 @@
  * _codecs_iso2022.c: Codecs collection for ISO-2022 encodings.
  *
  * Written by Hye-Shik Chang <perky@FreeBSD.org>
+ * $CJKCodecs: _codecs_iso2022.c,v 1.22 2004/08/19 17:08:13 perky Exp $
  */
 
 #define USING_IMPORTED_MAPS
@@ -103,7 +104,7 @@
 
 typedef int (*iso2022_init_func)(void);
 typedef ucs4_t (*iso2022_decode_func)(const unsigned char *data);
-typedef DBCHAR (*iso2022_encode_func)(const ucs4_t *data, Py_ssize_t *length);
+typedef DBCHAR (*iso2022_encode_func)(const ucs4_t *data, int *length);
 
 struct iso2022_designation {
 	unsigned char mark;
@@ -159,7 +160,7 @@ ENCODER(iso2022)
 		const struct iso2022_designation *dsg;
 		DBCHAR encoded;
 		ucs4_t c = **inbuf;
-		Py_ssize_t insize;
+		int insize;
 
 		if (c < 0x80) {
 			if (STATE_G0 != CHARSET_ASCII) {
@@ -182,7 +183,7 @@ ENCODER(iso2022)
 
 		encoded = MAP_UNMAPPABLE;
 		for (dsg = CONFIG_DESIGNATIONS; dsg->mark; dsg++) {
-			Py_ssize_t length = 1;
+			int length = 1;
 			encoded = dsg->encoder(&c, &length);
 			if (encoded == MAP_MULTIPLE_AVAIL) {
 				/* this implementation won't work for pair
@@ -299,12 +300,12 @@ DECODER_RESET(iso2022)
 	return 0;
 }
 
-static Py_ssize_t
+static int
 iso2022processesc(const void *config, MultibyteCodec_State *state,
-		  const unsigned char **inbuf, Py_ssize_t *inleft)
+		  const unsigned char **inbuf, size_t *inleft)
 {
 	unsigned char charset, designation;
-	Py_ssize_t i, esclen;
+	size_t i, esclen;
 
 	for (i = 1;i < MAX_ESCSEQLEN;i++) {
 		if (i >= *inleft)
@@ -387,10 +388,10 @@ iso2022processesc(const void *config, MultibyteCodec_State *state,
 	else if ((c) == 0xa2) (assi) = 0x2019;				\
 	else if ((c) == 0xaf) (assi) = 0x2015;
 
-static Py_ssize_t
+static int
 iso2022processg2(const void *config, MultibyteCodec_State *state,
-		 const unsigned char **inbuf, Py_ssize_t *inleft,
-		 Py_UNICODE **outbuf, Py_ssize_t *outleft)
+		 const unsigned char **inbuf, size_t *inleft,
+		 Py_UNICODE **outbuf, size_t *outleft)
 {
 	/* not written to use encoder, decoder functions because only few
 	 * encodings use G2 designations in CJKCodecs */
@@ -424,7 +425,7 @@ DECODER(iso2022)
 
 	while (inleft > 0) {
 		unsigned char c = IN1;
-		Py_ssize_t err;
+		int err;
 
 		if (STATE_GETFLAG(F_ESCTHROUGHOUT)) {
 			/* ESC throughout mode:
@@ -588,7 +589,7 @@ ksx1001_decoder(const unsigned char *data)
 }
 
 static DBCHAR
-ksx1001_encoder(const ucs4_t *data, Py_ssize_t *length)
+ksx1001_encoder(const ucs4_t *data, int *length)
 {
 	DBCHAR coded;
 	assert(*length == 1);
@@ -624,7 +625,7 @@ jisx0208_decoder(const unsigned char *data)
 }
 
 static DBCHAR
-jisx0208_encoder(const ucs4_t *data, Py_ssize_t *length)
+jisx0208_encoder(const ucs4_t *data, int *length)
 {
 	DBCHAR coded;
 	assert(*length == 1);
@@ -661,7 +662,7 @@ jisx0212_decoder(const unsigned char *data)
 }
 
 static DBCHAR
-jisx0212_encoder(const ucs4_t *data, Py_ssize_t *length)
+jisx0212_encoder(const ucs4_t *data, int *length)
 {
 	DBCHAR coded;
 	assert(*length == 1);
@@ -759,7 +760,7 @@ jisx0213_2004_2_decoder(const unsigned char *data)
 }
 
 static DBCHAR
-jisx0213_encoder(const ucs4_t *data, Py_ssize_t *length, void *config)
+jisx0213_encoder(const ucs4_t *data, int *length, void *config)
 {
 	DBCHAR coded;
 
@@ -813,7 +814,7 @@ jisx0213_encoder(const ucs4_t *data, Py_ssize_t *length, void *config)
 }
 
 static DBCHAR
-jisx0213_2000_1_encoder(const ucs4_t *data, Py_ssize_t *length)
+jisx0213_2000_1_encoder(const ucs4_t *data, int *length)
 {
 	DBCHAR coded = jisx0213_encoder(data, length, (void *)2000);
 	if (coded == MAP_UNMAPPABLE || coded == MAP_MULTIPLE_AVAIL)
@@ -825,10 +826,10 @@ jisx0213_2000_1_encoder(const ucs4_t *data, Py_ssize_t *length)
 }
 
 static DBCHAR
-jisx0213_2000_1_encoder_paironly(const ucs4_t *data, Py_ssize_t *length)
+jisx0213_2000_1_encoder_paironly(const ucs4_t *data, int *length)
 {
 	DBCHAR coded;
-	Py_ssize_t ilength = *length;
+	int ilength = *length;
 
 	coded = jisx0213_encoder(data, length, (void *)2000);
 	switch (ilength) {
@@ -848,7 +849,7 @@ jisx0213_2000_1_encoder_paironly(const ucs4_t *data, Py_ssize_t *length)
 }
 
 static DBCHAR
-jisx0213_2000_2_encoder(const ucs4_t *data, Py_ssize_t *length)
+jisx0213_2000_2_encoder(const ucs4_t *data, int *length)
 {
 	DBCHAR coded = jisx0213_encoder(data, length, (void *)2000);
 	if (coded == MAP_UNMAPPABLE || coded == MAP_MULTIPLE_AVAIL)
@@ -860,7 +861,7 @@ jisx0213_2000_2_encoder(const ucs4_t *data, Py_ssize_t *length)
 }
 
 static DBCHAR
-jisx0213_2004_1_encoder(const ucs4_t *data, Py_ssize_t *length)
+jisx0213_2004_1_encoder(const ucs4_t *data, int *length)
 {
 	DBCHAR coded = jisx0213_encoder(data, length, NULL);
 	if (coded == MAP_UNMAPPABLE || coded == MAP_MULTIPLE_AVAIL)
@@ -872,10 +873,10 @@ jisx0213_2004_1_encoder(const ucs4_t *data, Py_ssize_t *length)
 }
 
 static DBCHAR
-jisx0213_2004_1_encoder_paironly(const ucs4_t *data, Py_ssize_t *length)
+jisx0213_2004_1_encoder_paironly(const ucs4_t *data, int *length)
 {
 	DBCHAR coded;
-	Py_ssize_t ilength = *length;
+	int ilength = *length;
 
 	coded = jisx0213_encoder(data, length, NULL);
 	switch (ilength) {
@@ -895,7 +896,7 @@ jisx0213_2004_1_encoder_paironly(const ucs4_t *data, Py_ssize_t *length)
 }
 
 static DBCHAR
-jisx0213_2004_2_encoder(const ucs4_t *data, Py_ssize_t *length)
+jisx0213_2004_2_encoder(const ucs4_t *data, int *length)
 {
 	DBCHAR coded = jisx0213_encoder(data, length, NULL);
 	if (coded == MAP_UNMAPPABLE || coded == MAP_MULTIPLE_AVAIL)
@@ -916,7 +917,7 @@ jisx0201_r_decoder(const unsigned char *data)
 }
 
 static DBCHAR
-jisx0201_r_encoder(const ucs4_t *data, Py_ssize_t *length)
+jisx0201_r_encoder(const ucs4_t *data, int *length)
 {
 	DBCHAR coded;
 	JISX0201_R_ENCODE(*data, coded)
@@ -934,7 +935,7 @@ jisx0201_k_decoder(const unsigned char *data)
 }
 
 static DBCHAR
-jisx0201_k_encoder(const ucs4_t *data, Py_ssize_t *length)
+jisx0201_k_encoder(const ucs4_t *data, int *length)
 {
 	DBCHAR coded;
 	JISX0201_K_ENCODE(*data, coded)
@@ -966,7 +967,7 @@ gb2312_decoder(const unsigned char *data)
 }
 
 static DBCHAR
-gb2312_encoder(const ucs4_t *data, Py_ssize_t *length)
+gb2312_encoder(const ucs4_t *data, int *length)
 {
 	DBCHAR coded;
 	assert(*length == 1);
@@ -985,7 +986,7 @@ dummy_decoder(const unsigned char *data)
 }
 
 static DBCHAR
-dummy_encoder(const ucs4_t *data, Py_ssize_t *length)
+dummy_encoder(const ucs4_t *data, int *length)
 {
 	return MAP_UNMAPPABLE;
 }
