+++ Objects/sliceobject.c
@@ -80,10 +80,9 @@ PySlice_New(PyObject *start, PyObject *stop, PyObject *step)
 }
 
 int
-PySlice_GetIndices(PySliceObject *r, Py_ssize_t length,
-                   Py_ssize_t *start, Py_ssize_t *stop, Py_ssize_t *step)
+PySlice_GetIndices(PySliceObject *r, int length,
+                   int *start, int *stop, int *step)
 {
-	/* XXX support long ints */
 	if (r->step == Py_None) {
 		*step = 1;
 	} else {
@@ -111,12 +110,12 @@ PySlice_GetIndices(PySliceObject *r, Py_ssize_t length,
 }
 
 int
-PySlice_GetIndicesEx(PySliceObject *r, Py_ssize_t length,
-		     Py_ssize_t *start, Py_ssize_t *stop, Py_ssize_t *step, Py_ssize_t *slicelength)
+PySlice_GetIndicesEx(PySliceObject *r, int length,
+		     int *start, int *stop, int *step, int *slicelength)
 {
 	/* this is harder to get right than you might think */
 
-	Py_ssize_t defstart, defstop;
+	int defstart, defstop;
 
 	if (r->step == Py_None) {
 		*step = 1;
@@ -231,7 +230,7 @@ static PyMemberDef slice_members[] = {
 static PyObject*
 slice_indices(PySliceObject* self, PyObject* len)
 {
-	Py_ssize_t ilen, start, stop, step, slicelength;
+	int ilen, start, stop, step, slicelength;
 
 	ilen = PyInt_AsLong(len);
